refactor: begin migration to easynetworkmanager
- total project refactor and simplification
