+++ ESP/ini/dev_config.ini
@@ -20,8 +20,7 @@ lib_ldf_mode = deep+
 upload_speed = 921600
 lib_deps =
 	https://github.com/espressif/esp32-camera
-	https://github.com/me-no-dev/ESPAsyncWebServer.git
-	https://github.com/me-no-dev/AsyncTCP.git
+	https://github.com/ZanzyTHEbar/EasyNetworkManager.git
 extra_scripts =
 	pre:tools/customname.py
 	post:tools/createzip.py
