--- ESP/lib/src/data/StateManager/StateManager.cpp
@@ -1,10 +1,0 @@
-#include "StateManager.hpp"
-
-StateManager<State_e> stateManager;
-StateManager<WiFiState_e> wifiStateManager;
-StateManager<WebServerState_e> webServerStateManager;
-StateManager<MDNSState_e> mdnsStateManager;
-StateManager<CameraState_e> cameraStateManager;
-StateManager<LEDStates_e> ledStateManager;
-StateManager<StreamState_e> streamStateManager;
-StateManager<ConfigState_e> configStateManager;
