+++ ESP/lib/src/local/data/statemanager/state_manager.cpp
@@ -1,0 +1,6 @@
+#include "state_manager.hpp"
+
+StateManager<CameraState_e> cameraStateManager;
+StateManager<LEDStates_e> ledStateManager;
+StateManager<StreamState_e> streamStateManager;
+StateManager<ConfigState_e> configStateManager;
