+++ ESP/lib/src/local/io/camera/camera_handler.cpp
