+++ ESP/lib/src/local/logo/logo.hpp
